Split MAddressCommand into two clases
- One class for the command.
- One class for shared components.
